kubernetes-pulumi/pyproject.toml
[project]
name = "homelab-shared"
version = "0.1.0"
dependencies = [
    "pulumi>=3.0.0",
    "pulumi-kubernetes>=4.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "pulumiverse-doppler>=0.9.0",
    "pulumi-oci>=4.1.0",
    "pulumi-hcloud>=1.32.1",
    "pulumi-command>=1.2.0",
    "pulumi-random>=4.16.0",
    "pulumi-cloudflare>=5.12.0",
    "pulumi-authentik>=2024.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared"]
